refactor: move firebaseAuth under services dir
