+++ src/navigation/AppNavigator.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
-import { isUserLoggedIn } from '../screens/Auth/firebaseAuth';
+import { isUserLoggedIn } from '../services/firebaseAuth';
 
 // Import the navigators
 import AuthNavigator from './AuthNavigator';
